protocol helpers, beginning
git-svn-id: http://svn.miranda-ng.org/main/trunk@5285 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ protocols/Skype/src/skype_transfers.cpp
@@ -144,7 +144,7 @@ void CSkypeProto::OnFileEvent(const ConversationRef &conversation, const Message
 
 				PROTORECVFILET pre = {0};
 				pre.flags = PREF_TCHAR;
-				pre.fileCount = transfers.size();
+				pre.fileCount = (int)transfers.size();
 				pre.timestamp = timestamp;
 				pre.tszDescription = L"";
 				pre.ptszFiles = files;
